move the function to controller
